Module: Cleanup and refactor
