+++ natural_frontend/constants.py
@@ -3,7 +3,7 @@ FLASK = "Flask"
 
 FAST_API_ROUTE_TYPE = "APIRoute"
 
-BAD_ROUTES_BY_FRAMEWORK = {
-    FAST_API: ["handle_form", "frontend"],
+NON_API_ROUTES_BY_FRAMEWORK = {
+    FAST_API: [],
     FLASK: ["static"],
 }
